Separated all function declarations into separate files for a much more modular setup, and to save on compile time.
