+++ CMakeLists.txt
@@ -86,7 +86,6 @@ set(HEADER_FILES
 "include/defaults.h"
 "include/globals.h"
 "include/game.h"
-"include/utils.h"
 "include/sys_sprites.h"
 "include/sys_audio.h"
 "include/sys_wave_manager.h"
@@ -101,6 +100,16 @@ set(HEADER_FILES
 "include/data/data_sound.h"
 "include/data/data_enemy.h"
 "include/data/data_weapon.h"
+"include/utils_camera.h"
+"include/utils_sound.h"
+"include/utils_sprite.h"
+"include/utils_animation.h"
+"include/utils_ai_behavior.h"
+"include/utils_enemies.h"
+"include/utils_dynamic.h"
+"include/utils_stages.h"
+"include/utils_text.h"
+"include/utils_menu.h"
 "include/scene_title.h"
 "include/scene_menu.h"
 "include/scene_subweapon.h"
